index.jsx에서 App컴포넌트 분리
